レビュー投稿機能の実装

+++ db/migrate/20210105072806_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[6.0]
     create_table :reviews do |t|
       t.integer      :taste,              null: false
       t.integer      :scent,              null: false
-      t.integer      :recommend_score,    null: false
+      t.float        :recommend_score,    null: false
       t.string       :otsumami
       t.string       :comment,            null: false
       t.references   :user,               null: false
